requirements.txt
# MIDI & transcription
basic-pitch[tf]==0.4.0

# TensorFlow backend for Basic Pitch
tensorflow==2.15.0

# MIDI representation & audio synthesis
pretty_midi==0.2.11
pyfluidsynth==1.3.4

# WAV/MP3 handling
pydub==0.25.1
imageio-ffmpeg==0.6.0

# Numerical processing
numpy==1.26.4

# Optional utility
#typing_extensions==4.9.0

ollama~=0.6.1
simpleaudio~=1.0.4
librosa~=0.11.0
pip~=25.3
attrs~=25.4.0
setuptools~=60.2.0
mido~=1.3.1
packaging~=23.2
requests~=2.32.5
keras~=2.15.0
pyyaml~=6.0.3
cattrs~=25.3.0
llvmlite~=0.46.0
numba~=0.63.1
cffi~=2.0.0
scipy~=1.11.2
pooch~=1.8.2
tqdm~=4.67.1
platformdirs~=4.5.1
pyaml~=25.7.0
threadpoolctl~=3.6.0
mpmath~=1.3.0
sympy~=1.14.0
decorator~=5.2.1
msgpack~=1.1.2
joblib~=1.5.3
certifi~=2025.11.12
soundfile~=0.12.1
audioread~=3.1.0
soxr~=1.0.0
scikit-learn~=1.5.1
urllib3~=2.6.2
idna~=3.11
pycparser~=2.23
wheel~=0.37.1
coremltools~=9.0
onnxruntime~=1.16.1
resampy~=0.4.2
humanfriendly~=10.0
coloredlogs~=15.0.1
flatbuffers~=25.12.19
six~=1.17.0
fastapi
uvicorn
python-multipart
python-dotenv>=1.0.0

# Metrics and monitoring
prometheus_client>=0.17.0
psutil>=5.9.0